Model_Run_Many_Reps

Run SEIRV model for single region for large number of repetitions

Usage

Model_Run_Many_Reps(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  years_data = c(1940:1941),
  year0 = 1940,
  vaccine_efficacy = 1,
  time_inc = 1,
  output_type = "full",
  mode_start = 0,
  start_SEIRV = list(),
  mode_time = 0,
  n_reps = 1,
  division = 10,
  deterministic = FALSE
)

Arguments

FOI_spillover

Vector of values of force of infection due to spillover from sylvatic reservoir (size depends on mode_time)

R0

Vector of values of basic reproduction number for urban spread of infection (size depends on mode_time)

vacc_data

Projected vaccination-based immunity (assuming vaccine_efficacy = 1) by age group and year

pop_data

Population by age group and year

years_data

Incremental vector of years denoting years for which to save data

year0

First year in population/vaccination data

vaccine_efficacy

Proportional vaccine efficacy

time_inc

Time increment in days to use in model (should be 1.0, 2.5 or 5.0 days)model (should be 1.0, 2.5 or 5.0 days)

output_type

Type of data to output:
"full" = SEIRVC + FOI for all steps and ages
"infs" = annual total new infections (C) summed across all ages
"sero" = annual SEIRV
"infs_sero" = annual SEIRV, C summed across all ages
"infs_alt" = annual total new infections not combined by age
"infs_alt2" = total new infections combined by age for all steps

mode_start

Flag indicating how to set initial population immunity level in addition to vaccination
If mode_start = 0, only vaccinated individuals
If mode_start = 1, shift some non-vaccinated individuals into recovered to give herd immunity (stratified by age)
If mode_start = 2, use SEIRV input in list from previous run(s)

start_SEIRV

SEIRV data from end of a previous run to use as input (if mode_start = 2)

mode_time

Type of time dependence of FOI_spillover and R0 to be used:
If mode_time = 0, no time variation (constant values)
If mode_time = 1, FOI/R0 vary annually without seasonality (number of values = number of years to consider)
If mode_time = 2, FOI/R0 vary with monthly seasonality without inter - annual variation (number of values = 12)
If mode_time = 3, FOI/R0 vary with daily seasonality without inter - annual variation (number of values = 365/dt)
If mode_time = 4, FOI/R0 vary annually with monthly seasonality (number of values = 12*number of years to consider)
If mode_time = 5, FOI/R0 vary annually with daily seasonality (number of values = (365/dt)*number of years to consider)

n_reps

Number of repetitions (used to set number of particles and threads)

division

Number of particles/threads to run in one go (up to 20)

deterministic

TRUE/FALSE - set model to run in deterministic mode if TRUE '

Details

Accepts epidemiological + population parameters and model settings; runs SEIRV model for one region over a specified time period for a number of repetitions and outputs time-dependent SEIRV values, infection numbers and/or total force of infection values. Variation of Model_Run() used for running a large number of repetitions (>20).