case_data_calculate_multi

Calculate reported case data from SEIRV model output across multiple repetitions

Usage

case_data_calculate_multi(
  model_data = list(),
  p_severe_inf = 0.12,
  p_death_severe_inf = 0.39,
  p_rep_severe = 1,
  p_rep_death = 1,
  output_type = "annual",
  deterministic = FALSE
)

Arguments

model_data

SEIRV output of Model_Run and similar functions

p_severe_inf

Probability of an infection being severe

p_death_severe_inf

Probability of a severe infection resulting in death

p_rep_severe

Probability of reporting of a severe but non-fatal infection

p_rep_death

Probability of reporting of a fatal infection

output_type

Type of output to produce: "annual" - Total reported cases and reported deaths by year "pts" - Total reported cases and reported deaths for every time point, summed over age groups "full" - Reported cases and reported deaths for every time point and age group

deterministic

Indicates whether to calculate results deterministically (TRUE) or stochastically (FALSE) '

Details

[TBA]