create_param_labels

Apply names to the parameters in a set used for data matching and parameter fitting

Usage

create_param_labels(enviro_data = NULL, extra_estimated_params = c("vacc_eff"))

Arguments

enviro_data

Environmental data frame, containing only relevant environmental covariate values for regions of interest

extra_estimated_params

Vector of strings listing variable parameters besides ones determining FOI/R0 (may include vaccine efficacy and/or infection/death reporting probabilities and/or Brazil FOI adjustment factor)

Details

Takes in environmental covariate data along with names of additional parameters (vaccine efficacy and reporting probabilities) and generates list of names for parameter set to use as input for fitting functions