input_data_process

Cross-reference input data with serological and/or annual case/death data for comparison

Usage

input_data_process(
  input_data = list(),
  obs_sero_data = NULL,
  obs_case_data = NULL
)

Arguments

input_data

List of population and vaccination data for multiple regions (created using create_input_data() function and usually loaded from an RDS file)

obs_sero_data

Seroprevalence data for comparison, by region, year & age group, in format no. samples/no. positives

obs_case_data

Annual reported case/death data for comparison, by region and year, in format no. cases/no. deaths

Details

This function, used to prepare input data for functions used to calculate the likelihood of observed data, amends a list of population and vaccination data used as input for other functions, cross-referencing it with seroprevalence and/or case data and adding connection information.