Parameter setup

Set up parameters to input into model

Usage

parameter_setup(
  FOI_spillover = 0,
  R0 = 1,
  vacc_data = list(),
  pop_data = list(),
  year0 = 1940,
  years_data = c(1941:1942),
  mode_start = 0,
  vaccine_efficacy = 1,
  start_SEIRV = list(),
  dt = 1
)

Arguments

FOI_spillover

Force of infection due to spillover from sylvatic reservoir

R0

Reproduction number for urban spread of infection

vacc_data

Projected vaccination-based immunity (assuming vaccine_efficacy = 1) by age group and year

pop_data

Population by age group and year

year0

First year in population/vaccination data

years_data

Incremental vector of years denoting years for which to save data

mode_start

Flag indicating how to set initial population immunity level in addition to vaccination If mode_start = 0, only vaccinated individuals If mode_start = 1, shift some non-vaccinated individuals into recovered to give herd immunity (uniform by age, R0 based only) If mode_start = 2, use SEIRV input in list from previous run(s) If mode_start = 3, shift some non-vaccinated individuals into recovered to give herd immunity (stratified by age)

vaccine_efficacy

Proportional vaccine efficacy

start_SEIRV

SEIRV data from end of a previous run to use as input

dt

Time increment in days to use in model (should be either 1.0 or 5.0 days) '

Details

Takes in multiple inputs, outputs list for use by odin SEIRV model.