single_posterior_calc

Function which calculates and outputs posterior likelihood of observing simulated data

Usage

single_posterior_calc(
  log_params_prop = c(),
  input_data = list(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  ...
)

Arguments

log_params_prop

Proposed values of parameters to be estimated (natural logarithm of actual parameters)

input_data

List of population and vaccination data for multiple regions (created using data input creation code and usually loaded from RDS file), with cross-reference tables added in MCMC

obs_sero_data

Seroprevalence data for comparison, by region, year & age group, in format no. samples/no. positives

obs_case_data

Annual reported case/death data for comparison, by region and year, in format no. cases/no. deaths

...

= Constant parameters/flags/etc. loaded to or determined by mcmc() and mcmc_prelim_fit, including params_data, mode_start, time_inc, n_reps, enviro_data, add_est_param_names, deterministic, mode_time, mode_parallel, cluster, i_FOI_const, i_FOI_var, i_R0_const, i_R0_var, i_FOI_prior, i_R0_prior

Details

This function calculates the posterior likelihood of observing a set of observations (across multiple regions and data types) for a given proposed parameter set. [TBA]