mcmc_checks

Perform checks on MCMC inputs

Usage

mcmc_checks(
  log_params_ini = c(),
  n_regions = 1,
  prior_settings = list(type = "zero"),
  enviro_data = list(),
  add_values = list(vaccine_efficacy = 1, p_rep_severe = 1, p_rep_death = 1, m_FOI_Brazil
    = 1),
  extra_estimated_params = list()
)

Arguments

log_params_ini

Initial values of parameters to be estimated (natural logarithm of actual parameters)

n_regions

Number of regions

prior_settings

List containing settings for priors: must contain text named "type":
If type = "zero", prior probability is always zero
If type = "flat", prior probability is zero if log parameter values in designated ranges log_params_min and log_params_max, -Inf otherwise; log_params_min and log_params_max included in prior_settings as vectors of same length as log_params_ini
If type = "norm", prior probability is given by dnorm calculation on parameter values with settings based on vectors of values in prior_settings:
norm_params_mean and norm_params_sd (vectors of mean and standard deviation values applied to log FOI/R0 parameters and to actual values of additional parameters)
+ FOI_mean + FOI_sd (mean + standard deviation of computed FOI, single values)
+ R0_mean + R0_sd (mean + standard deviation of computed R0, single values)

enviro_data

Values of environmental covariates

add_values

List of parameters in addition to those governing FOI/R0, either giving a fixed value or giving NA to indicate that they are part of the parameter set to be estimated
vaccine_efficacy Vaccine efficacy (proportion of reported vaccinations causing immunity) (must be present)
p_rep_severe Probability of observation of severe infection
p_rep_death Probability of observation of death
m_FOI_Brazil Multiplier of spillover FOI for Brazil regions (only relevant if regions in Brazil to be considered)

extra_estimated_params

Vector of names of parameters to be estimated in addition to those governing FOI and R0; see add_values

Details

This function, which is called by MCMC(), performs a number of checks on data to be used in fitting to ensure proper functionality. It verifies that the number of parameters being estimated is consistent with other settings and that certain values are not outwith sensible boundaries (e.g. probabilities must be between 0 and 1).